Rename ShortGraphQLApi to ShortLinkGraphQLApi
