+++ frontend/src/service/ShortLink.service.ts
@@ -1,6 +1,6 @@
 import { Url } from '../entity/Url';
 import { Err, ErrorService } from './Error.service';
-import { ShortGraphQLApi } from './ShortGraphQL.api';
+import { ShortLinkGraphQLApi } from './ShortLinkGraphQL.api';
 
 export interface IPagedShortLinks {
   shortLinks: Url[];
@@ -11,7 +11,7 @@ export interface IPagedShortLinks {
 
 export class ShortLinkService {
   constructor(
-    private shortGraphQLApi: ShortGraphQLApi,
+    private shortLinkGraphQLApiService: ShortLinkGraphQLApi,
     private errorService: ErrorService
   ) {}
 
@@ -21,7 +21,7 @@ export class ShortLinkService {
   ): Promise<IPagedShortLinks> {
     return new Promise((resolve, reject) => {
       // TODO(issue#673): support pagination for user created Short Links in API.
-      this.shortGraphQLApi
+      this.shortLinkGraphQLApiService
         .getUserShortLinks(offset, pageSize)
         .then((URLs: Url[]) => {
           resolve(this.getPagedShortLinks(URLs, offset, pageSize));
